requirements.txt
streamlit
cohere
pinecone-client
pypdf
python-dotenv
